+++ spec/whois/record/parser/responses/whois.srs.net.nz/status_registered_spec.rb
@@ -21,6 +21,16 @@ describe Whois::Record::Parser::WhoisSrsNetNz, "status_registered.expected" do
     described_class.new(part)
   end
 
+  describe "#domain" do
+    it do
+      subject.domain.should == "google.co.nz"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      lambda { subject.domain_id }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
   describe "#status" do
     it do
       subject.status.should == :registered
@@ -52,6 +62,81 @@ describe Whois::Record::Parser::WhoisSrsNetNz, "status_registered.expected" do
     it do
       subject.expires_on.should be_a(Time)
       subject.expires_on.should == Time.parse("2014-02-17 00:00:00 +13:00")
+    end
+  end
+  describe "#registrar" do
+    it do
+      subject.registrar.should be_a(Whois::Record::Registrar)
+      subject.registrar.id.should           == nil
+      subject.registrar.name.should         == "MarkMonitor"
+      subject.registrar.organization.should == nil
+      subject.registrar.url.should          == nil
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      subject.registrant_contacts.should be_a(Array)
+      subject.registrant_contacts.should have(1).items
+      subject.registrant_contacts[0].should be_a(Whois::Record::Contact)
+      subject.registrant_contacts[0].type.should          == Whois::Record::Contact::TYPE_REGISTRANT
+      subject.registrant_contacts[0].id.should            == nil
+      subject.registrant_contacts[0].name.should          == "Google Inc"
+      subject.registrant_contacts[0].organization.should  == nil
+      subject.registrant_contacts[0].address.should       == "1600 Amphitheatre Parkway"
+      subject.registrant_contacts[0].city.should          == "Mountain View"
+      subject.registrant_contacts[0].zip.should           == "94043"
+      subject.registrant_contacts[0].state.should         == "CA"
+      subject.registrant_contacts[0].country.should       == "US (UNITED STATES)"
+      subject.registrant_contacts[0].country_code.should  == nil
+      subject.registrant_contacts[0].phone.should         == "+1 650 +1 650 3300100"
+      subject.registrant_contacts[0].fax.should           == "+1 650 +1 650 6181434"
+      subject.registrant_contacts[0].email.should         == "dns-admin@google.com"
+      subject.registrant_contacts[0].created_on.should    == nil
+      subject.registrant_contacts[0].updated_on.should    == nil
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      subject.admin_contacts.should be_a(Array)
+      subject.admin_contacts.should have(1).items
+      subject.admin_contacts[0].should be_a(Whois::Record::Contact)
+      subject.admin_contacts[0].type.should          == Whois::Record::Contact::TYPE_ADMIN
+      subject.admin_contacts[0].id.should            == nil
+      subject.admin_contacts[0].name.should          == "Google Inc"
+      subject.admin_contacts[0].organization.should  == nil
+      subject.admin_contacts[0].address.should       == "1600 Amphitheatre Parkway"
+      subject.admin_contacts[0].city.should          == "Mountain View"
+      subject.admin_contacts[0].zip.should           == "94043"
+      subject.admin_contacts[0].state.should         == "CA"
+      subject.admin_contacts[0].country.should       == "US (UNITED STATES)"
+      subject.admin_contacts[0].country_code.should  == nil
+      subject.admin_contacts[0].phone.should         == "+1 650 +1 650 3300100"
+      subject.admin_contacts[0].fax.should           == "+1 650 +1 650 6181434"
+      subject.admin_contacts[0].email.should         == "dns-admin@google.com"
+      subject.admin_contacts[0].created_on.should    == nil
+      subject.admin_contacts[0].updated_on.should    == nil
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      subject.technical_contacts.should be_a(Array)
+      subject.technical_contacts.should have(1).items
+      subject.technical_contacts[0].should be_a(Whois::Record::Contact)
+      subject.technical_contacts[0].type.should          == Whois::Record::Contact::TYPE_TECHNICAL
+      subject.technical_contacts[0].id.should            == nil
+      subject.technical_contacts[0].name.should          == "Google Inc"
+      subject.technical_contacts[0].organization.should  == nil
+      subject.technical_contacts[0].address.should       == "1600 Amphitheatre Parkway"
+      subject.technical_contacts[0].city.should          == "Mountain View"
+      subject.technical_contacts[0].zip.should           == "94043"
+      subject.technical_contacts[0].state.should         == "CA"
+      subject.technical_contacts[0].country.should       == "US (UNITED STATES)"
+      subject.technical_contacts[0].country_code.should  == nil
+      subject.technical_contacts[0].phone.should         == "+1 650 +1 650 3300100"
+      subject.technical_contacts[0].fax.should           == "+  +1 650 6181434"
+      subject.technical_contacts[0].email.should         == "dns-admin@google.com"
+      subject.technical_contacts[0].created_on.should    == nil
+      subject.technical_contacts[0].updated_on.should    == nil
     end
   end
   describe "#nameservers" do