+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - NEW: Added #domain_id to CoCCA parsers.
 
+- NEW: Added full whois.srs.net.nz parser.
+
 - CHANGED: Updated whois.nic.ac, whois.nic.io, and whois.nic.sh parsers to the new response format (GH-238).
 
 - CHANGED: Updated whois1.nic.bi parser to the new response format.